WIP Move data out of JS, display labels without JS, shrink chart.js.
chart.js before was 149Kb (41Kb gzipped) plus you needed moment
(50Kb/17Kb), total of 199Kb/58Kb. By stripping out unused graphs and
code and config options, and switching to category axis rather than
time, it is now a total of 39Kb/13Kb.

The data needs to be in the template so we can generate it on the
server.